feat: 팔로워, 팔로잉 삭제시 숫자 연동 (#278)
