+++ frontend/carbook/src/pages/Profile/Profile.ts
@@ -85,6 +85,31 @@ export default class ProfilePage extends Component {
     `;
   }
 
+  async deleteFollower(nickname: string) {
+    await basicAPI.delete(`/api/profile/follower?follower=${nickname}`);
+    this.receiveFollowLists(this.state.profileMode);
+    this.fetchProfilePage(this.state.nickname);
+  }
+
+  async deleteFollowing(nickname: string) {
+    await basicAPI.post("/api/profile/follow", { followingNickname: nickname });
+    this.receiveFollowLists(this.state.profileMode);
+    this.fetchProfilePage(this.state.nickname);
+  }
+
+  async receiveFollowLists(profileMode: string) {
+    const mode = profileMode === "follower" ? "followers" : "followings";
+    const data = await basicAPI
+      .get(`/api/profile/${mode}?nickname=${this.state.nickname}`)
+      .then((response) => response.data)
+      .catch((error) => error);
+
+    this.setState({
+      ...this.state,
+      follows: data.nicknames,
+    });
+  }
+
   render(): void {
     if (this.state.isloading || this.state?.notSession) return;
 
@@ -113,15 +138,35 @@ export default class ProfilePage extends Component {
       const followingSection = target.closest("section.profile-following");
       const followButton = target.closest(".follow-button");
       const modifyInfoButton = target.closest(".modify-button");
+      const deleteButton = target.closest(".follower-delete-button");
+
+      if (deleteButton) {
+        const nickname = (
+          this.$target.querySelector(".follower-info") as HTMLElement
+        ).dataset.nickname as string;
+        const mode = this.$target.querySelector(
+          ".profile__contents-header"
+        )?.innerHTML;
+
+        mode === "팔로워"
+          ? this.deleteFollower(nickname)
+          : this.deleteFollowing(nickname);
+        return;
+      }
+
       if (postsSection) {
         this.setState({ ...this.state, profileMode: "posts" });
         return;
       }
       if (followerSection) {
+        const profileMode = "follower";
+        this.receiveFollowLists(profileMode);
         this.setState({ ...this.state, profileMode: "follower" });
         return;
       }
       if (followingSection) {
+        const profileMode = "following";
+        this.receiveFollowLists(profileMode);
         this.setState({ ...this.state, profileMode: "following" });
         return;
       }
@@ -151,8 +196,6 @@ export default class ProfilePage extends Component {
         showErrorModal(modal, "회원정보가 변경되었습니다");
       }
     });
-
-    this.setEvent();
     this.mounted();
   }
 
@@ -188,7 +231,7 @@ export default class ProfilePage extends Component {
       new Followlists(profile_contents, {
         profileMode: this.state.profileMode,
         isMyProfile: this.state.isMyProfile,
-        follows: this.state.followers,
+        follows: this.state.follows,
         nickname: this.state.nickname,
       });
 
@@ -196,7 +239,7 @@ export default class ProfilePage extends Component {
       new Followlists(profile_contents, {
         profileMode: this.state.profileMode,
         isMyProfile: this.state.isMyProfile,
-        follows: this.state.followings,
+        follows: this.state.follows,
         nickname: this.state.nickname,
       });
